requirements1.txt
streamlit
pandas
openai
python-dotenv
matplotlib
fpdf
